времянка

+++ src/main/java/ru/job4j/serialization/json/Main.java
@@ -6,7 +6,7 @@ import com.google.gson.GsonBuilder;
 public class Main {
     public static void main(String[] args) {
         final Student student = new Student(37, true, "Alex Pozharov",
-                new Subject("Maths", new byte[] {5, 4, 5, 5}));
+                new Subject("Maths", new int[] {5, 4, 5, 5}));
         final Gson gson = new GsonBuilder().create();
         System.out.println(gson.toJson(student));
         final String studentJson =
